New Sophia Branch.
Mostly just cleaned up the scripts to look better, I tried not to change anything. The only thing I changed was Sam's pan_tilt.py script where I removed the confidence threshold stuff, which can be used in the main script.

Otherwise the main.py is the script that will control OOOoOoO all the other scripts. The pan_preview.py is the script to test the live preview/yolo camera with sams pan and tilt.